reduce element type cases for compile times

+++ test/test_expect.jl
@@ -17,7 +17,7 @@ using Test: @testset, @test
         (named_grid((nx, ny)), "square"),
     ]
     for (g, g_str) in gs
-        ψ = random_tensornetworkstate(ComplexF32, g, "S=1/2"; bond_dimension = χ)
+        ψ = random_tensornetworkstate(ComplexF64, g, "S=1/2"; bond_dimension = χ)
         v_centre = first(center(g))
 
         sz_exact = expect(ψ, ("Z", v_centre); alg = "exact")
@@ -32,7 +32,7 @@ using Test: @testset, @test
         Rmps = 16
         sz_boundarymps = expect(ψ, ("Z", v_centre); alg = "boundarymps", mps_bond_dimension = Rmps)
 
-        @test sz_boundarymps ≈ sz_exact atol = 10*eps(Float32)
+        @test sz_boundarymps ≈ sz_exact atol = 10*eps(Float64)
 
         if !is_tree(g)
             v_centre_neighbor = first(neighbors(g, v_centre))
